feat: add dashboard API
- add sales summary
- add most sold products
- add most sold categories
- add total item and total quantity in expense and income

+++ database/migrations/2024_04_30_020159_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->float('price', 10, 2);
-            $table->tinyInteger('quantity');
+            $table->integer('quantity');
             $table->foreignId('created_by')->constrained('users');
             $table->timestamps();
             $table->softDeletes();
